+++ app/src/features/surveys/telemetry/ManualTelemetryPage.tsx
@@ -3,7 +3,7 @@ import CircularProgress from '@mui/material/CircularProgress';
 import { grey } from '@mui/material/colors';
 import Paper from '@mui/material/Paper';
 import { SurveyContext } from 'contexts/surveyContext';
-import { TelemetryContextProvider } from 'contexts/telemetryContext';
+import { TelemetryContextProvider } from 'contexts/telemetryDataContext';
 import { useContext } from 'react';
 import SurveyObservationHeader from '../observations/SurveyObservationHeader';
 import ManualTelemetryComponent from './ManualTelemetryComponent';
